+++ Scripts/Git/push.sh
@@ -7,27 +7,28 @@ if [ ! -d .git ]; then
 fi
 
 # Get current directory path and file name
-DIR=$(pwd)
-CHANGELOG="$DIR/CHANGELOG.md"
+dir=$(pwd)
+changelog="$dir/CHANGELOG.md"
+version="$dir/VERSION.md"
 
 # Save YAML header to file
 # Run git log and save the output to file
 # Add YAML closure to the end of the file
-echo '```' > "$CHANGELOG"
-git log --graph --pretty=format:'Commit: %h - %s%nAuthor: %an - %cD%n' --decorate >> "$CHANGELOG"
-echo '```' >> "$CHANGELOG"
+echo '```' > "$changelog"
+git log --graph --pretty=format:'Commit: %h - %s%nAuthor: %an - %cD%n' --decorate >> "$changelog"
+echo '```' >> "$changelog"
 
 # Show confirmation message
 echo "Changelog successfully generated at: ./CHANGELOG.md"
 
 # Check if VERSION.md file exists
-if [ ! -f "VERSION.md" ]; then
+if [ ! -f "$version" ]; then
     echo "Error: The VERSION.md file is not found. Aborting."
     exit 1
 fi
 
 # Read the contents of VERSION.md file and remove the backticks (`)
-version_number=$(cat VERSION.md | sed 's/`//g')
+version_number=$(cat "$version" | sed 's/`//g')
 
 # Check that VERSION.md file is not empty
 if [ -z "$version_number" ]; then
@@ -41,7 +42,8 @@ read -p "Are you sure you want to push this commit? [Y/n]: " confirmation
 
 # Check the user's response
 if [ "$confirmation" != "Y" ]; then
-    echo "Operation cancelled. Neither commit nor push will be performed."
+    echo "Operation cancelled."
+    echo "Neither commit nor push will be performed."
     exit 0
 fi