+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 ```
+* Commit: 240d5f3 - v196
+| Author: fjrodafo - Wed, 5 Feb 2025 14:46:01 +0000
+| 
 * Commit: 990a8e5 - v195
 | Author: fjrodafo - Wed, 5 Feb 2025 14:43:58 +0000
 |